+++ PsychSourceGL/Source/Windows/Base/PsychTimeGlue.c
@@ -21,15 +21,8 @@
  *
  */
 
-// Must include initguid.h before Psych.h, and therefore before any windows includes, or
-// use of any GUID as part of power management setup won't work (build time link failure):
-#include <initguid.h>
-
 // Regular includes:
 #include "Psych.h"
-
-// For power management control:
-#include <PowrProf.h>
 
 /*
  *        file local state variables
@@ -1540,37 +1533,6 @@ void PsychSetThreadName(const char *name)
     // No-op for now.
     (void) name;
     return;
-}
-
-/* Disable (dontIdle = TRUE), or enable use of processor core idling by OS processor power management,
- * return if mode switch worked (TRUE) or failed (FALSE). Disabling idling should, e.g., prevent the
- * cpus from entering sleep states, ie. ACPI C states greater than C0. As C state transitions, especially
- * from deeper C states, can be time consuming, they can cause significant execution timing jitter in the
- * above 200 microseconds range, and thereby impair timing correctness tests, timing calibrations etc.
- *
- * Note that this will increase cpu utilization to permanent 100% on all cores and may cause drastic
- * increase in power consumption and heat production! It should therefore generally not be used all
- * the time, but only for short time intervals, e.g., during startup tests and calibrations.
- *
- * This functionality requires Windows Vista and later, or build time failure will happen.
- */
-psych_bool PsychOSDisableProcessorIdling(psych_bool dontIdle)
-{
-    GUID* activeScheme;
-
-    if (PowerGetActiveScheme(NULL, &activeScheme) ||
-        PowerWriteACValueIndex(NULL, activeScheme, &GUID_PROCESSOR_SETTINGS_SUBGROUP, &GUID_PROCESSOR_IDLE_DISABLE, dontIdle ? 1 : 0) ||
-        PowerWriteDCValueIndex(NULL, activeScheme, &GUID_PROCESSOR_SETTINGS_SUBGROUP, &GUID_PROCESSOR_IDLE_DISABLE, dontIdle ? 1 : 0) ||
-        PowerSetActiveScheme(NULL, activeScheme)) {
-        // Failed to switch processor idle mode!
-        return(FALSE);
-    }
-
-    // Give processor power management a bit of time to latch into the new state:
-    PsychYieldIntervalSeconds(0.010);
-
-    // Success:
-    return(TRUE);
 }
 
 /* Initialize condition variable:
